requirements.txt
# MONAI i PyTorch
torch==2.1.0
torchvision==0.16.0
monai==1.3.0

# Data processing
numpy==1.24.3
Pillow==10.1.0
opencv-python==4.8.1.78
scikit-learn==1.3.2

# Visualization
matplotlib==3.8.0
tqdm==4.66.1

# API (jeśli chcesz Flask z PyTorch)
Flask==3.0.0
flask-cors==4.0.0

# Optional - dla ONNX export
onnx==1.15.0
onnxruntime==1.16.3